+++ Libs/YahnisElsts/PluginUpdateChecker/Puc/v5p2/Metadata.php
@@ -1,5 +1,5 @@
 <?php
-namespace YahnisElsts\PluginUpdateChecker\v5p2;
+namespace WordPress\Plugins\PpWpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5p2;
 
 use LogicException;
 use stdClass;