+++ Libs/YahnisElsts/PluginUpdateChecker/Puc/v5p2/Plugin/PluginInfo.php
@@ -1,7 +1,7 @@
 <?php
-namespace YahnisElsts\PluginUpdateChecker\v5p2\Plugin;
+namespace WordPress\Plugins\PpWpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5p2\Plugin;
 
-use YahnisElsts\PluginUpdateChecker\v5p2\Metadata;
+use WordPress\Plugins\PpWpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5p2\Metadata;
 
 if ( !class_exists(PluginInfo::class, false) ):