+++ Libs/YahnisElsts/PluginUpdateChecker/Puc/v5p2/Plugin/Ui.php
@@ -1,5 +1,5 @@
 <?php
-namespace YahnisElsts\PluginUpdateChecker\v5p2\Plugin;
+namespace WordPress\Plugins\PpWpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5p2\Plugin;
 
 if ( !class_exists('Ui', false) ):
 	/**