+++ Libs/YahnisElsts/PluginUpdateChecker/Puc/v5p2/Vcs/BaseChecker.php
@@ -1,5 +1,5 @@
 <?php
-namespace YahnisElsts\PluginUpdateChecker\v5p2\Vcs;
+namespace WordPress\Plugins\PpWpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5p2\Vcs;
 
 if ( !interface_exists(BaseChecker::class, false) ):