+++ Libs/YahnisElsts/PluginUpdateChecker/Puc/v5p2/Vcs/ReleaseAssetSupport.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace YahnisElsts\PluginUpdateChecker\v5p2\Vcs;
+namespace WordPress\Plugins\PpWpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5p2\Vcs;
 
 if ( !trait_exists(ReleaseAssetSupport::class, false) ) :