+++ Libs/YahnisElsts/PluginUpdateChecker/Puc/v5p2/Vcs/ThemeUpdateChecker.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace YahnisElsts\PluginUpdateChecker\v5p2\Vcs;
+namespace WordPress\Plugins\PpWpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5p2\Vcs;
 
-use YahnisElsts\PluginUpdateChecker\v5p2\Theme;
-use YahnisElsts\PluginUpdateChecker\v5p2\Utils;
+use WordPress\Plugins\PpWpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5p2\Theme;
+use WordPress\Plugins\PpWpMemoryUsage\Libs\YahnisElsts\PluginUpdateChecker\v5p2\Utils;
 
 if ( !class_exists(ThemeUpdateChecker::class, false) ):