+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## \[In Development\] - Unreleased
 
+## \[1.2.0\] - 2023-09-15
+
+### Changed
+
+- Integrate PluginUpdateChecker into our namespace
+
+### Removed
+
+- Redundant argument
+
 ## \[1.1.0\] - 2023-09-12
 
 ### Added